bill-easy-frontend.com/bill-easy-backend.com/backend/requirements.txt
# FastAPI framework
fastapi==0.115.2

# ASGI server for FastAPI
uvicorn[standard]==0.30.6

# Database and ORM
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# Data validation and serialization
pydantic==2.8.2
pydantic-settings==2.3.4

# Migrations
alembic==1.13.2

# Environment variables
python-dotenv==1.0.1

# Password hashing and authentication
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# Date/time handling
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.2
httpx==0.27.0

# Optional utilities
requests==2.32.3

